move odf plotting to external program
