+++ qsiprep/niworkflows/viz/utils.py
@@ -17,15 +17,11 @@ import numpy as np
 import nibabel as nb
 import matplotlib.pyplot as plt
 
-from dipy.core.ndindex import ndindex
-from dipy.reconst.odf import gfa
-from dipy.direction import peak_directions
 from lxml import etree
 from nilearn import image as nlimage
 from nilearn.plotting import plot_anat
 from svgutils.transform import SVGFigure
 from seaborn import color_palette
-from PIL import Image
 
 from .. import NIWORKFLOWS_LOG
 from nipype.utils import filemanip
@@ -789,264 +785,3 @@ def slices_from_bbox(mask_data, cuts=3, padding=0):
         slices = [start + (i + 1) * inc for i in range(cuts + 2 * padding)]
         vox_coords.append(slices[padding:-padding])
     return {k: [int(_v) for _v in v] for k, v in zip(['x', 'y', 'z'], vox_coords)}
-
-
-def plot_peak_slice(odf_4d, sphere, background_data, out_file, axis, slicenum, mask_data,
-                    tile_size=1200, normalize_peaks=True):
-    from fury import actor, window
-    view_up = [(0., 0., 1.), (0., 0., 1.), (0., -1., 0.)]
-
-    # Make a slice mask to reduce memory
-    new_shape = list(odf_4d.shape)
-    new_shape[axis] = 1
-    image_shape = new_shape[:3]
-    midpoint = (new_shape[0] / 2., new_shape[1] / 2., new_shape[2] / 2.)
-
-    if axis == 0:
-        odf_slice = odf_4d[slicenum, :, :, :].reshape(new_shape)
-        image_slice = background_data[slicenum, :, :].reshape(image_shape)
-        mask_slice = mask_data[slicenum, :, :].reshape(image_shape)
-        camera_dist = max(midpoint[1], midpoint[2]) * np.pi
-    elif axis == 1:
-        odf_slice = odf_4d[:, slicenum, :, :].reshape(new_shape)
-        image_slice = background_data[:, slicenum, :].reshape(image_shape)
-        mask_slice = mask_data[:, slicenum, :].reshape(image_shape)
-        camera_dist = max(midpoint[0], midpoint[2]) * np.pi
-    elif axis == 2:
-        odf_slice = odf_4d[:, :, slicenum, :].reshape(new_shape)
-        image_slice = background_data[:, :, slicenum].reshape(image_shape)
-        mask_slice = mask_data[:, :, slicenum].reshape(image_shape)
-        camera_dist = max(midpoint[0], midpoint[1]) * np.pi
-    position = list(midpoint)
-    position[axis] += camera_dist
-
-    # Find the actual peaks
-    peak_dirs, peak_values = peaks_from_odfs(odf_slice, sphere,
-                                             relative_peak_threshold=.1,
-                                             min_separation_angle=15,
-                                             mask=mask_slice,
-                                             normalize_peaks=normalize_peaks,
-                                             npeaks=3)
-    if normalize_peaks:
-        peak_values = peak_values / peak_values.max() * np.pi
-    peak_actor = actor.peak_slicer(peak_dirs, peak_values, colors=None)
-    image_actor = actor.slicer(image_slice, opacity=0.6, interpolation='nearest')
-    image_size = (tile_size, tile_size)
-    scene = window.Scene()
-    scene.add(image_actor)
-    scene.add(peak_actor)
-
-    xfov_min, xfov_max = 0, new_shape[0] - 1
-    yfov_min, yfov_max = 0, new_shape[1] - 1
-    zfov_min, zfov_max = 0, new_shape[2] - 1
-    peak_actor.display_extent(xfov_min, xfov_max, yfov_min, yfov_max, zfov_min, zfov_max)
-    image_actor.display_extent(xfov_min, xfov_max, yfov_min, yfov_max, zfov_min, zfov_max)
-    scene.set_camera(focal_point=tuple(midpoint),
-                     position=tuple(position),
-                     view_up=view_up[axis])
-    window.record(scene, out_path=out_file, reset_camera=False, size=image_size)
-    scene.clear()
-    del peak_actor, image_actor, scene
-
-
-def peak_slice_series(odf_4d, sphere, background_data, out_file, mask_image=None,
-                      prefix='odf', tile_size=1200, n_cuts=3, padding=4,
-                      normalize_peaks=True):
-
-    # Make a slice mask to reduce memory
-    if mask_image is None:
-        image_mask = np.ones(background_data.shape)
-    else:
-        image_mask = nb.load(mask_image).get_fdata()
-
-    slice_indices = slices_from_bbox(background_data, cuts=n_cuts, padding=padding)
-    # Render the axial slices
-    z_image = Image.new('RGB', (tile_size, tile_size * n_cuts))
-    for slicenum, z_slice in enumerate(slice_indices['z']):
-        png_file = '{}_tra_{:03d}.png'.format(prefix, z_slice)
-        plot_peak_slice(odf_4d, sphere, background_data, png_file, 2, z_slice, image_mask,
-                        tile_size, normalize_peaks)
-        z_image.paste(Image.open(png_file), (0, slicenum * tile_size))
-
-    # Render the sagittal slices
-    x_image = Image.new('RGB', (tile_size, tile_size * n_cuts))
-    for slicenum, x_slice in enumerate(slice_indices['x']):
-        png_file = '{}_sag_{:03d}.png'.format(prefix, x_slice)
-        plot_peak_slice(odf_4d, sphere, background_data, png_file, 0, x_slice, image_mask,
-                        tile_size, normalize_peaks)
-        x_image.paste(Image.open(png_file), (0, slicenum * tile_size))
-
-    # Render the coronal slices
-    y_image = Image.new('RGB', (tile_size, tile_size * n_cuts))
-    for slicenum, y_slice in enumerate(slice_indices['y']):
-        png_file = '{}_cor_{:03d}.png'.format(prefix, y_slice)
-        plot_peak_slice(odf_4d, sphere, background_data, png_file, 1, y_slice, image_mask,
-                        tile_size, normalize_peaks)
-        y_image.paste(Image.open(png_file), (0, slicenum * tile_size))
-
-    final_image = Image.new('RGB', (tile_size * 3, tile_size * n_cuts))
-    final_image.paste(z_image, (0, 0))
-    final_image.paste(x_image, (tile_size, 0))
-    final_image.paste(y_image, (tile_size * 2, 0))
-    final_image.save(out_file)
-
-
-def peaks_from_odfs(odf4d, sphere, relative_peak_threshold,
-                    min_separation_angle, mask=None,
-                    gfa_thr=0, normalize_peaks=False,
-                    npeaks=5):
-
-    shape = odf4d.shape[:-1]
-    if mask is None:
-        mask = np.ones(shape, dtype='bool')
-    else:
-        if mask.shape != shape:
-            raise ValueError("Mask is not the same shape as data.")
-
-    gfa_array = np.zeros(shape)
-    qa_array = np.zeros((shape + (npeaks,)))
-
-    peak_dirs = np.zeros((shape + (npeaks, 3)))
-    peak_values = np.zeros((shape + (npeaks,)))
-    peak_indices = np.zeros((shape + (npeaks,)), dtype='int')
-    peak_indices.fill(-1)
-
-    global_max = -np.inf
-    for idx in ndindex(shape):
-        if not mask[idx]:
-            continue
-        odf = odf4d[idx]
-        gfa_array[idx] = gfa(odf)
-        if gfa_array[idx] < gfa_thr:
-            global_max = max(global_max, odf.max())
-            continue
-        # Get peaks of odf
-        direction, pk, ind = peak_directions(odf, sphere,
-                                             relative_peak_threshold,
-                                             min_separation_angle)
-        # Calculate peak metrics
-        if pk.shape[0] != 0:
-            global_max = max(global_max, pk[0])
-            n = min(npeaks, pk.shape[0])
-            qa_array[idx][:n] = pk[:n] - odf.min()
-            peak_dirs[idx][:n] = direction[:n]
-            peak_indices[idx][:n] = ind[:n]
-            peak_values[idx][:n] = pk[:n]
-
-            if normalize_peaks:
-                peak_values[idx][:n] /= pk[0]
-                peak_dirs[idx] *= peak_values[idx][:, None]
-
-    qa_array /= global_max
-
-    return peak_dirs, peak_values
-
-
-def get_camera_for_roi(roi_data, roi_id, view_axis):
-    voxel_coords = np.row_stack(np.nonzero(roi_data == roi_id))
-    centroid = voxel_coords.mean(1)
-    other_axes = [[1, 2], [0, 2], [0, 1]][view_axis]
-    projected_x = voxel_coords[other_axes[0]]
-    projected_y = voxel_coords[other_axes[1]]
-    xspan = projected_x.max() - projected_x.min()
-    yspan = projected_y.max() - projected_y.min()
-    camera_distance = max(xspan, yspan) * np.pi
-    return centroid, camera_distance
-
-
-def plot_an_odf_slice(odf_4d, full_sphere, background_data, tile_size, filename,
-                      centroid, axis, camera_distance, subtract_iso, mask_image):
-    from fury import actor, window
-    view_up = [(0., 0., 1.), (0., 0., 1.), (0., -1., 0.)]
-
-    # Adjust the centroid so it's only a single slice
-    slicenum = int(np.round(centroid)[axis])
-    centroid[axis] = 0
-    position = centroid.copy()
-    position[axis] = position[axis] + camera_distance
-
-    # Roll if viewing an axial slice
-    roll = 3 if axis == 2 else 0
-    position[1] = position[1] - roll
-
-    # Ensure the dimensions reflect that there is only one slice
-    new_shape = list(odf_4d.shape)
-    new_shape[axis] = 1
-    image_shape = new_shape[:3]
-    if axis == 0:
-        odf_slice = odf_4d[slicenum, :, :, :].reshape(new_shape)
-        image_slice = background_data[slicenum, :, :].reshape(image_shape)
-    elif axis == 1:
-        odf_slice = odf_4d[:, slicenum, :, :].reshape(new_shape)
-        image_slice = background_data[:, slicenum, :].reshape(image_shape)
-    elif axis == 2:
-        odf_slice = odf_4d[:, :, slicenum, :].reshape(new_shape)
-        image_slice = background_data[:, :, slicenum].reshape(image_shape)
-
-    # Tile to get the whole ODF
-    odf_slice = np.tile(odf_slice, (1, 1, 1, 2))
-    if subtract_iso:
-        odf_slice = odf_slice - odf_slice.min(3, keepdims=True)
-    # Make graphics objects
-    odf_actor = actor.odf_slicer(odf_slice, sphere=full_sphere,
-                                 colormap=None, scale=0.6, mask=image_slice)
-    image_actor = actor.slicer(image_slice, opacity=0.6, interpolation='nearest')
-    image_size = (tile_size, tile_size)
-    scene = window.Scene()
-    scene.add(image_actor)
-    scene.add(odf_actor)
-    xfov_min, xfov_max = 0, new_shape[0] - 1
-    yfov_min, yfov_max = 0, new_shape[1] - 1
-    zfov_min, zfov_max = 0, new_shape[2] - 1
-    odf_actor.display_extent(xfov_min, xfov_max, yfov_min, yfov_max, zfov_min, zfov_max)
-    image_actor.display_extent(xfov_min, xfov_max, yfov_min, yfov_max, zfov_min, zfov_max)
-    scene.set_camera(focal_point=tuple(centroid),
-                     position=tuple(position),
-                     view_up=view_up[axis])
-    window.record(scene, out_path=filename, reset_camera=False, size=image_size)
-    scene.clear()
-    del odf_actor, image_actor, scene
-
-
-def odf_roi_plot(odf_4d, halfsphere, background_data, out_file, roi_file,
-                 prefix='odf', tile_size=1200, subtract_iso=False, mask=None):
-
-    roi_data = nb.load(roi_file).get_fdata()
-    roi_image = Image.new('RGB', (tile_size * 3, tile_size))
-    roi1_centroid, roi1_distance = get_camera_for_roi(roi_data, 1, 2)
-    roi2_centroid, roi2_distance = get_camera_for_roi(roi_data, 2, 1)
-    roi3_centroid, roi3_distance = get_camera_for_roi(roi_data, 3, 1)
-
-    camera_distance = max(roi1_distance, roi2_distance, roi3_distance)
-    # Make a slice mask to reduce memory
-    if mask is None:
-        image_mask = np.ones(roi_data.shape)
-    else:
-        image_mask = nb.load(mask).get_fdata()
-
-    # Fill out the other half of the sphere
-    odf_sphere = halfsphere.mirror()
-    semiovale_axial_file = '{}_semoivale_axial.png'.format(prefix)
-    plot_an_odf_slice(odf_4d, odf_sphere, background_data, tile_size,
-                      semiovale_axial_file, centroid=roi1_centroid, axis=2,
-                      camera_distance=camera_distance, subtract_iso=subtract_iso,
-                      mask_image=image_mask)
-    roi_image.paste(Image.open(semiovale_axial_file), (0, 0))
-
-    # Render the coronal slice with a double-crossing
-    cst_x_cc_file = '{}_CSTxCC.png'.format(prefix)
-    plot_an_odf_slice(odf_4d, odf_sphere, background_data, tile_size,
-                      cst_x_cc_file, centroid=roi2_centroid, axis=1,
-                      camera_distance=camera_distance, subtract_iso=subtract_iso,
-                      mask_image=image_mask)
-    roi_image.paste(Image.open(cst_x_cc_file), (tile_size, 0))
-
-    # Render the corpus callosum
-    cc_file = '{}_CC.png'.format(prefix)
-    plot_an_odf_slice(odf_4d, odf_sphere, background_data, tile_size,
-                      cc_file, centroid=roi3_centroid, axis=1,
-                      camera_distance=camera_distance, subtract_iso=subtract_iso,
-                      mask_image=image_mask)
-    roi_image.paste(Image.open(cc_file), (2 * tile_size, 0))
-
-    roi_image.save(out_file)
