+++ setup.cfg
@@ -117,6 +117,7 @@ console_scripts =
     mif2fib=qsiprep.cli.convertODFs:mif_to_fib
     fib2mif=qsiprep.cli.convertODFs:fib_to_mif
     qsiprep_group_report=qsiprep.cli.group_report:aggregate_reports
+    recon_plot=qsiprep.cli.recon_plot:recon_plot
 
 
 [flake8]
